migrated wbot initialization to typescript
