+++ backend/src/@types/WAWebJS.d.ts
@@ -1,0 +1,5 @@
+declare namespace WAWebJS {
+  export interface ClientInfo {
+    meuId: number;
+  }
+}
