+++ backend/src/@types/express.d.ts
@@ -1,5 +1,5 @@
 declare namespace Express {
-	export interface Request {
-		user: { id: string; profile: string };
-	}
+  export interface Request {
+    user: { id: string; profile: string };
+  }
 }
