+++ backend/src/@types/qrcode-terminal.d.ts
@@ -1,0 +1,1 @@
+declare module "qrcode-terminal";
